requirements.txt
pydantic>=2.0
databricks-sdk>=0.28.0
databricks-connect>=14.3
databricks-vectorsearch>=0.40
langchain-databricks>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
pandas>=2.0
uvicorn>=0.20.0,<1
PyYAML>=6.0.1
mlflow-skinny>=2.11.3
